! move client based functions to their own directory browser detector is a stretch here, but what the heck
